Player Manager Implementation
Player manager now handles player health and ammo. Ammo types are defined as an enum in the player manager class, and the manager uses the event system to update the values when actions like reloading and shooting occur. You can also get the current ammo count of any type using a function call.